requirements.txt
streamlit==1.31.0
openai==0.28.1
python-dotenv==1.0.0
sqlalchemy==2.0.27
bcrypt==4.1.2
pandas==2.2.0
plotly==5.18.0 